Adds PINOperationQueue
PINOperationQueue is an operation queue designed to allow
specifying the maximum number of threads, support priority (ordering,
not thread priority and cancelation).

It is also designed to prevent live locks due to priority inversion.
It attempts to achieve this by maintaining two queues: a serial queue
which always takes operations in FIFO order and a concurrent queue
which runs operations in priority order.